requirements.txt
numpy
msgpack-numpy
lmdb

hydra-core
pytorch-lightning

./pointnet2_ops_lib/.
